Fixed a crash/major bug in the original C# code
As far as I can tell, the code was reading either 16 or 32 bytes
(depending on the edition of the hash) into a buffer, and had some logic
to handle cases where an update wsa triggered with less than a full
buffer, in which case the state would contain how many bits were
successfully read and the function would return to the caller without
actually updating anything.

The code was using the offset from the last call instead of from the
beginning of the state buffer. Sometimes there would not be enough bytes
left (which would also imply that the initial bytes were not fed into
the hasher). I'm not sure how the old code ever passed any test, ever.
It just simply didn't work.

We're probably going to do away with the Wilhelm Liao's code entirely,
this has really eliminated any baseline confidence in the correctness of
the initial C# port.